Consolidate fused-LLM ops under agilerl/algorithms/core/llm_ops/
PR review feedback from mikepratt1:

1. ``llm_policy_loss.py:23`` — "Move to ``fused_llm_policy_loss.py``"
2. ``fused_llm_policy_loss.py:32`` — "Combine with other fused files in
   ``llm_ops/`` directory"

Restructure:

  agilerl/algorithms/core/
  ├── fused_llm_policy_loss.py   (removed)
  ├── fused_lora.py              (removed)
  ├── llm_policy_loss.py         (removed)
  └── llm_ops/
      ├── __init__.py            (re-exports public surface)
      ├── fused_dpo_loss.py      (the ``_LigerDPOWithAlpha`` autograd
      │                          Function, previously in
      │                          ``agilerl/utils/llm_utils.py``)
      ├── fused_lora.py          (moved)
      └── fused_policy_loss.py   (merged llm_policy_loss.py +
                                  fused_llm_policy_loss.py — plain-tensor
                                  ``_k3_kl`` / ``llm_policy_loss_fn``
                                  helpers at top, Liger autograd Function
                                  below)

Behavioural change: ``fused_policy_loss`` no longer raises at import
when ``liger-kernel`` is missing — instead it stubs the
``LigerFusedLinearPPOBase`` to ``object`` so the class body resolves,
then reassigns ``LigerFusedLinearPolicyLossFunction = None`` at the
bottom. This lets ``LLMAlgorithm._calculate_kl_divergence`` (in
``base.py``) import ``_k3_kl`` from the consolidated module on
platforms without Liger, while keeping the existing ``is None`` guard
pattern for the autograd Function.

Local: 358 tests pass on macOS (no Liger).
